Increments Epoch to 1
Found a bug in our nightly build system after the introduction of the
new testing build suffixes where the latest builds weren't actually
being counted as the latest builds.

As seen in:
```
18.04.0~ce~dev~git20180315.170650.0.8fabfd2-0~debian
18.04.0~ce~dev~git20180314.170605.0.8024288-0~debian
...
18.04.0~ce~dev~3~git20180319.170823.0.4836f8c-0~debian
```

Where the last image is actually the latest package but is counted as
being a lesser version due to the inclusion of the `3` in the package
name.

Signed-off-by: Eli Uriegas <eli.uriegas@docker.com>