Set kernel.pid_max sysctl to a bigger value
Default for kernel.pid_max sysctl appears to be
128K (i.e. 131072), which might not be enough for
systems with lots of RAM running lots of threads.

Once the limit is hit, bad things might happen. In particular,
Go's runtime/cgo can't handle it gracefully, chosing
a suicide instead:

> runtime/cgo: pthread_create failed: Resource temporarily unavailable
> SIGABRT: abort

What needs to be done is to carefully audit dockerd for goroutine
leaks. In the meantime, though, raising the pid_max limit seems
to be a not-so-bad workaround.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>