CLI RPM: only conflict with older engine packages
Engine packages 18.09 and up no longer contain the CLI (as it is
split to a separate package), so we can limit the conflict to
older packages.

This change should allow users to install a Docker CE CLI on
a host that has Docker EE Engine 18.09, or vice-versa (an
EE CLI on a host that has Docker CE Engine 18.09).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>