Makefile: rename DEB_VERSION, STATIC_VERSION to GEN_*_VER
DEB_VERSION and STATIC_VERSION are a macro (evaluated/expanded when
called), so renaming them makes it more clear that they're expanded
at that point. GEN_*_VER is also in line with the equivalent
GEN_RPM_VERSION in the rpm/Makefile.

The environment variable used in the container (and scripts) is
kept as DEB_VERSION.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>