deb: add apparmor as "recommends" on Debian as well
Debian did not have AppArmor available until Debian 7 "Wheezy", and not have it
enabled by _default_ until Debian 10 "Buster". The packaging scripts did not
add AppArmor as recommended dependency for that reason.

Now that Debian 10 "Buster" is the current stable, and older releases reached
EOL, we can remove the special handling for Debian/Ubuntu, and unconditionally
add apparmor as a recommended dependency.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>