buildx: remove "-docker" suffix from version
When building the CLI plugin, we were adding a "-docker" suffix. It's a bit
unclear if there was a specific reason for this (possibly to differentiate
between builds from GitHub and builds that are part of the CLI?

Initially, the version had a `-tp-docker` prefix (where `-tp` was used to
indicate it was still experimental); this prefix was removed in commit
0b45cce1abcf8868f4d957317a839f1e80058607

    /usr/libexec/docker/cli-plugins/docker-buildx docker-cli-plugin-metadata
    {
         "SchemaVersion": "0.1.0",
         "Vendor": "Docker Inc.",
         "Version": "v0.5.1-docker",
         "ShortDescription": "Build with BuildKit"
    }

This patch removes the `-docker` suffix from the version string.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>