rpm: docker-compose-plugin: allow stand-alone installation
This removes the "require" docker-ce-cli, so that the package can be installed
stand-alone (withouth the docker cli itself installed).

Unlike "deb" packages, rpm package do not have "soft" requirement ("enhances",
"recommends"), so having a "requires" makes the docker-cli a hard requirement.

While in most situations, the plugin would be installed together with the docker
cli, the plugin _is_ able to run standalone:

    /usr/libexec/docker/cli-plugins/docker-compose up
    [+] Running 2/2
     ⠿ Network test_default  Created                         4.3s
     ⠿ Container test_web_1  Started                         2.1s
    Attaching to web_1
    web_1  | 10-listen-on-ipv6-by-default.sh: Getting the checksum of /etc/nginx/conf.d/default.conf
    web_1  | 10-listen-on-ipv6-by-default.sh: Enabled listen on IPv6 in /etc/nginx/conf.d/default.conf
    web_1  | /docker-entrypoint.sh: Launching /docker-entrypoint.d/20-envsubst-on-templates.sh
    web_1  | /docker-entrypoint.sh: Configuration complete; ready for start up

This removes the "Requires" annotation from the package, to simplify the dependency
graph, and to allow installing the package separately (which could be for testing
purposes, or for a minimal install).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit cab7b567fdd3e7108af65ed2734820d3095d2e4d)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>