rpm: docker-ce-cli: add docker-compose-plugin as "required"
Unfortunately (unlike deb packages), rpm packages do not have "soft" dependencies,
so to define a relation between the docker cli and the compose-plugin, we must
use "requires".

Alternatively, we can omit the dependency, to make the plugin an "optional" install,
and handle this through documentation, and in the script at download.docker.com.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>