static: add option to specify containerd and runc version
This allows passing `CONTAINERD_VERSION` and `RUNC_VERSION` make vars
to override the default version in the upstream repository's Dockerfile.

With this, it's easier to make the static packages match the latest
released `containerd.io` deb/rpm (which are used by the .deb and .rpm
docker-ce packages), without having to modify the upstream moby repository.

Currently, this uses the DOCKER_BUILD_OPTS make variable, which is available
in the moby Makefile, but work is in progress to add `CONTAINERD_VERSION`
and `RUNC_VERSION` make variables in the upstream repository. Once those
changes are merged, we can update the makefile in this repository accordingly.

With this patch:

If `RUNC_VERSION` and `CONTAINERD_VERSION` are not passed, the defaults are used:

    $ make \
        DOCKER_BUILD_PKGS=static-linux \
        REF=v20.10.8 \
        VERSION=v20.10.8 \
        static

    $ docker run --rm -v $(pwd)/static/build/linux/docker/:/docker alpine sh -c '/docker/containerd --version && /docker/runc --version'
    containerd github.com/containerd/containerd v1.4.9 e25210fe30a0a703442421b0f60afac609f950a3
    runc version 1.0.1
    commit: v1.0.1-0-g4144b638
    spec: 1.0.2-dev
    go: go1.16.8
    libseccomp: 2.4.4

Passing the `RUNC_VERSION` and `CONTAINERD_VERSION` vars overrides the version
of containerd and runc:

    $ make \
        DOCKER_BUILD_PKGS=static-linux \
        REF=v20.10.8 \
        VERSION=v20.10.8 \
        RUNC_VERSION=v1.0.2 \
        CONTAINERD_VERSION=v1.4.10 \
        static

    $ docker run --rm -v $(pwd)/static/build/linux/docker/:/docker alpine sh -c '/docker/containerd --version && /docker/runc --version'
    containerd github.com/containerd/containerd v1.4.10 8848fdb7c4ae3815afcc990a8a99d663dda1b590
    runc version 1.0.2
    commit: v1.0.2-0-g52b36a2d
    spec: 1.0.2-dev
    go: go1.16.8
    libseccomp: 2.4.4

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit f0302b8ab4656f38cb398a018719ca34ef3eabc9)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>