rpm/centos-8: adjust SUITE variable to keep it the same as before
This was a bit of an  oversight in 49ff9113494665d7aff0496a6aacf54a96bc344a;
when checking if the `$SUITE` variable was used elsewhere, I searched for
`$SUITE`, and not for `${SUITE}`. While it looks like the variable is not
used for any of the rpm-related build-scripts, it _is_ used in deb-related
scripts.

Let's change it back to the previous value; just in case it will be used
somewhere and causing side-effects.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>