Jenkinsfile: use quay.io/centos/centos:stream8 as well
The `image` field in the Jenkinsfile is not really used in this repository
(other than to name the stage), but in our internal pipeline, the equivalent
is used for some validation steps (to be upstreamed to this repository).

Let's update the Jenkinsfile to match our internal one, and to make sure
we don't forget updating it when we upstream some of those validation steps.

Relates to 49ff9113494665d7aff0496a6aacf54a96bc344a

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>