rpmlint: fix "W: no-documentation"
These packages were causing a warning that they didn't contain documentation.
For some of these, we need to add man-pages, but at least we can add some of
the files we have. Also adding `AUTHORS` and `SECURITY.md` to existing packages:

    docker-ce.x86_64: W: no-documentation
    docker-ce-rootless-extras.x86_64: W: no-documentation
    docker-compose-plugin.x86_64: W: no-documentation
    docker-scan-plugin.x86_64: W: no-documentation
    The package contains no documentation (README, doc, etc). You have to include
    documentation files.

I'm using unique directory names for each package to collect the docs in, as
we build the packages in the same environment (so possibly they would trip over
each-other), but possibly this is not needed (so "just to be sure").

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>