verify: rpm: don't perform dnf upgrade
This was added as part of 6dfaf9cc43198e7f0d0f87b7bcc1d78525662882, which came
from our internal repositories, but the original commits from there didn't have
context why an upgrade was performed.

In general, it's best practice to not perform a (dist-)upgrade inside the image,
and just depend on the image maintainers to have the image being up-to-date, so
removing this step.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>