deb: move packaging revision before distro information
Commit f8299f20216a4e2def617f7d3605eab1d68132ee introduced the `$pkgRevision`
in the package version, but we now actually had two separate escape hatches
for when we needed to do a packaging-only update; one before the packaging
version, and one after the distro-ID (VERSION_ID). The latter one was confusing,
as it appeared to be part of the distro-version (e.g. `22.04.0`).

This patch removes the additional version, and moves it to the start, and updates
the default to use "1", which aligns with the "_release" used for RPM packages,
and may help with warnings that the package version is not an "upstream" (debian)
version.

Comparing same version old vs new (works because we now use 1 as default):

    dpkg --compare-versions "23.0.0~rc.1-1~ubuntu.22.04~jammy" ">>" "23.0.0~rc.1-0~ubuntu.22.04.0~jammy" && echo "OK" || echo "KO"
    OK

Comparing newer version (new format) vs older version (old format) is ok:

    dpkg --compare-versions "23.0.0~rc.2-1~ubuntu.22.04~jammy" ">>" "23.0.0~rc.1-0~ubuntu.22.04.0~jammy" && echo "OK" || echo "KO"
    OK

Comparing same version, but newer $pkgRevision (either new format <-> old-format,
or new-format <-> new-format) is also ok:

    dpkg --compare-versions "23.0.0~rc.1-2~ubuntu.22.04~jammy" ">>" "23.0.0~rc.1-1~ubuntu.22.04~jammy" && echo "OK" || echo "KO"
    OK

    dpkg --compare-versions "23.0.0~rc.1-2~ubuntu.22.04~jammy" ">>" "23.0.0~rc.1-1~ubuntu.22.04.0~jammy" && echo "OK" || echo "KO"
    OK

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>