Remove libbtrfs-dev, btrfs-tools, btrfs-progs-devel build-dependencies
Starting with https://github.com/moby/moby/commit/6834304febca3c0610d1061bd9612ee99020df34,
building the engine no longer requires these headers, as it now uses the
kernel UAPI, which means we no longer need these packages as dependency.

Note that the static builds may still need them for containerd, but
dependencies for static builds are defined in the Dockerfile (containerd
is also in the process of migrating to use the kernel UAPI).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>