deb, rpm: add /etc/docker directory
Previous versions of the engine created this directory as a side-effect of the
(legacy) "key.json" file. With the removal of libtrust (and the key.json) file,
that directory is no longer created.

While the precence of this directory is not needed for the daemon to function,
users may expect it to be there, so it there should be no harm in creating it.

For scripting purposes, users are still recommended to check if the directory
exists or to create it (mkdir -p).

This patch adds a .dirs file to create the directory on installation for .deb;
https://www.debian.org/doc/manuals/maint-guide/dother.en.html#dirs

And adds a %dirs directive for .rpm packages:
http://ftp.rpm.org/max-rpm/s1-rpm-inside-files-list-directives.html#S3-RPM-INSIDE-DIR-DIRECTIVE

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>