rpm: recommend, not require  "docker-ce-rootless-extras" where possible
The deb pacakges have this dependency as "recommends", but older versions
of yum (RHEL7 / CentOS 7) do not support this. and only support "requires".

This patch uses a similar approach as bb4bd31ab60fe5bc496021d2606882db913c5bad
did for the CLI, and changes the dependency to be "recommends" where possible,
falling back to "requires" for older rpm versions.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>