static: remove unused docker-proxy.exe binary from Windows packages
The `docker-proxy.exe` binary was added to the static Windows packages through
commit 09541b553c2f7646626efcc2797bd92ee48a66d0 (which introduced a bug, later
fixed through fc5379fee869e8fa0e9a2569a66083e61e83de18).

It looks like that commit added the `docker-proxy.exe` binary with the
assumption that it's used on Windows (given that its also built in upstream),
but recent discussions revealed that's not the case.

This patch removes binary from the static packages for Windows, because it's
not used.

A quick look at Docker Desktop's build scripts show that it's not included
in those packages, and installation instructions for static binaries on
Windows do not mention this binary, so no further changes should be needed.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>