rpm: fix missing GOTOOLCHAIN=local in centos, rhel Dockerfiles
Commit a4090a0e198517b9e7603db375a283c3df3ec4f7 added GOTOOLCHAIN=local,
but for some reason missed the centos Dockerfiles.

The env-var is set to make sure we don't get unexpected updates of the go
toolchain when building. We need to set this env-var, because we're not using
the official golang image as base-image, but instead copy the binaries into
a distro-image.

This patch adds the missing GOTOOLCHAIN env-vars.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit dc175c09fd68aee12ca0752f83aff16a7a51fef2)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>