deb,rpm: fix Dockerfile default value for ARG linting warnings
The `GO_IMAGE` build-arg doesn't have a default;

     - InvalidDefaultArgInFrom: Default value for ARG ${GO_IMAGE} results in empty or invalid base image name (line 8)

This is not a real issue when building using the Makefiles (which is how
these Dockerfiles are expected to be used) but we can probably set a
default.

This patch sets the default to `golang:latest` to pick the latest version,
which should be fine for this (until we get a linter recommending not
to use `:latest`).

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>
(cherry picked from commit 9c73210773887ab31680a96542425dcf88fc2329)
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>