rpm: remove rpmlint from Dockerfiles as it's not used
Commit 784a53cec5061e6a718acc3a7104c61215c15078 added `rpmlint` to the
Dockerfiles and added a step in the Dockerfile to validate the SPEC files.
The validation step was removed in c245ce12c1e0a092ef21094b287b8cb94a51ee1e,
but didn't remove the step to install the package.

The `rpmlint` package is also being removed from CentOS Stream 10 ("CS10");

- https://issues.redhat.com/browse/CS-2451
- https://issues.redhat.com/browse/CS-2453

Remove the package as it's not used currently.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>