deb, rpm: remove libseccomp-dev, libseccomp2 dependencies
We used to have `libseccomp-dev` as a build-time dependency, but this
part is handled by the OCI runtime (runc). Similarly, the  `libseccomp2`
runtime dependency is used by `runc`, not by `dockerd` directly.

The docker-ce package depends on `containerd.io`, which defines `libseccomp2`
as a dependency, so we can remove it from the docker-ce package;

    apt-cache depends containerd.io
    containerd.io
      Depends: libc6
      Depends: libseccomp2
      Conflicts: containerd
      Conflicts: runc
      Replaces: containerd
        containerd.io
      Replaces: runc
        containerd.io

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>