deb: stop using dh_bash-completion for bash completions
The dh_bash-completion debhelper provides an easy way to install the
shell-completion scripts for Bash. Unfortunately there is no stable
equivalent yet for the other shells (zsh, fish, powershell), which
resulted in two out of three shells requiring manual install.

Given that the installation path for Bash is [well-documented][1],
we can align Bash with the other shells to make this less confusing.

This patch makes that change, and adds a code-comment to outline
the reasoning (and possible future options) for future readers.

[1]: https://github.com/scop/bash-completion/blob/79fd051907328c8c26372691f68d627c1f0e3916/README.md

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>