rpm: adjust spec to account for varying list of man-page sections
The CLI currently builds man-pages for sections 1, 5, and 8, but this
list may change; for example, the dockerd man-page is currently built
from the CLI repository but intended to be removed.

This patch changes the code to check all (1..9) secions, and install the
man-pages found in them.

This patch also updates the %doc and %files section; from the Fedora packaging
guide (see [1]);

> The %files section holds a list of all the files that RPM should install
> from the package. **This list should be exhaustive**, so that the RPM system
> knows exactly what your package installs. There are some options, though,
> to name all the files within a directory to help with packages containing
> hundreds of files.
> (...)
> In addition to naming each file on a line, you can use glob-style wildcards.

But also worth noting that it's NOT needed to mark manpages as %doc, see [2];

> Note also that files installed in %{_mandir} are automatically marked by RPM
> as documentation. Thus it is not necessary to use %doc.

So this patch:

- uses wildcards to enumerate all manpages in all sections
- removes the %doc for manpages, as this is automatic.

[1]: https://jfearn.fedorapeople.org/en-US/RPM/4/html/RPM_Guide/ch09s05.html
[2]: https://docs.fedoraproject.org/en-US/packaging-guidelines/#_manpages

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>