RHEL RPM: require nftables
The dynamically linked RHEL dockerd doesn't strictly need
a dependency on nftables (because it doesn't link against
libnftables.so). So, commit 1a73c6b didn't make it one.

But, it'll already be installed in most places anyway, it'll
be required when we deprecate iptables support, and it's
needed to try out the experimental nftables support.

So, make it a hard dependency now.

Signed-off-by: Rob Murray <rob.murray@docker.com>